Create room page
Room page represents the main screen of the project. This is where users are going to interact.
Room page is implemented with displaying instants. Instants are icons which will be shown for a short time to instantly react to something.

This implementation is only frontend-wise. There's no connection to the backend yet.